Integrated Booking and View with DB
Integrated GUI with DB except for where Business seats are to be booked as the DB Function required for that is missing. Moreover, there are some issues with the View Booking, i.e., the type and seats are not returned for the Booking. This is because when a booking is created, the bookingId in the seats documents are not updated, and therefore, no seat is associated with a bookingId.
The current changes are ready to be merged, however, I will prefer that first the above mentioned issues are resolved.