Added Popups for different messages
Popups can now be displayed for different messages and types, such as error, information, warning, questions, etc.
For now, the popups are only used for Cancel Booking and will be added for all other screens. Check if all the functionality is working on the front end and ONLY then merge.